Pull proxy_unavailable_services out into its own module
